session3 homework

+++ session3/q7.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 
 void main() {
   print('enter any number');
-  int number = int.parse(stdin.readLineSync()!);
+  double number = double.parse(stdin.readLineSync()!);
   if (number < 0) {
     print('this is negative number');
   } else if (number > 0) {
